pyproject.toml
[project]
name = "ai-drawer"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai (>=1.93.0,<2.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "python-telegram-bot (>=22.1,<23.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
